- Regenerate projects with the latest vapor templates

+++ vapor-mysql/projects/ItWorks-MySQL/Tests/AppTests/PostControllerTests.swift
@@ -22,13 +22,21 @@ class PostControllerTests: TestCase {
     let controller = PostController()
     
     func testPostRoutes() throws {        
-        let idOne = try create() ?? -1
+        guard let postOne = try storeNewPost(), let idOne = postOne.id?.int else {
+            XCTFail()
+            return
+        }
+
         try fetchOne(id: idOne)
         try fetchAll(expectCount: 1)
         try patch(id: idOne)
         try put(id: idOne)
 
-        let idTwo = try create() ?? -1
+        guard let postTwo = try storeNewPost(), let idTwo = postTwo.id?.int else {
+            XCTFail()
+            return
+        }
+
         try fetchAll(expectCount: 2)
 
         try deleteOne(id: idOne)
@@ -37,29 +45,32 @@ class PostControllerTests: TestCase {
         try deleteOne(id: idTwo)
         try fetchAll(expectCount: 0)
 
-        let newIds = try (1...5).map { _ in try create() ?? 1 }
-        try fetchAll(expectCount: newIds.count)
+        for _ in 1...5 {
+            _ = try storeNewPost()
+        }
+        try fetchAll(expectCount: 5)
         try deleteAll()
         try fetchAll(expectCount: 0)
     }
 
-    func create() throws -> Int? {
+    func storeNewPost() throws -> Post? {
         let req = Request.makeTest(method: .post)
         req.json = try JSON(node: ["content": initialMessage])
-        let res = try controller.create(request: req).makeResponse()
+        let res = try controller.store(req).makeResponse()
         
         let json = res.json
         XCTAssertNotNil(json)
+        let newId: Int? = try json?.get("id")
+        XCTAssertNotNil(newId)
         XCTAssertNotNil(json?["content"])
-        XCTAssertNotNil(json?["id"])
         XCTAssertEqual(json?["content"], req.json?["content"])
-        return try json?.get("id")
+        return try Post.find(newId)
     }
 
     func fetchOne(id: Int) throws {
         let req = Request.makeTest(method: .get)
         let post = try Post.find(id)!
-        let res = try controller.show(request: req, post: post).makeResponse()
+        let res = try controller.show(req, post: post).makeResponse()
 
         let json = res.json
         XCTAssertNotNil(json)
@@ -71,7 +82,7 @@ class PostControllerTests: TestCase {
 
     func fetchAll(expectCount count: Int) throws {
         let req = Request.makeTest(method: .get)
-        let res = try controller.index(request: req).makeResponse()
+        let res = try controller.index(req).makeResponse()
 
         let json = res.json
         XCTAssertNotNil(json?.array)
@@ -82,7 +93,7 @@ class PostControllerTests: TestCase {
         let req = Request.makeTest(method: .patch)
         req.json = try JSON(node: ["content": updatedMessage])
         let post = try Post.find(id)!
-        let res = try controller.update(request: req, post: post).makeResponse()
+        let res = try controller.update(req, post: post).makeResponse()
         
         let json = res.json
         XCTAssertNotNil(json)
@@ -96,7 +107,7 @@ class PostControllerTests: TestCase {
         let req = Request.makeTest(method: .put)
         req.json = try JSON(node: ["content": updatedMessage])
         let post = try Post.find(id)!
-        let res = try controller.replace(request: req, post: post).makeResponse()
+        let res = try controller.replace(req, post: post).makeResponse()
 
         let json = res.json
         XCTAssertNotNil(json)
@@ -110,12 +121,12 @@ class PostControllerTests: TestCase {
         let req = Request.makeTest(method: .delete)
         
         let post = try Post.find(id)!
-        _ = try controller.delete(request: req, post: post)
+        _ = try controller.delete(req, post: post)
     }
 
     func deleteAll() throws {
         let req = Request.makeTest(method: .delete)
-        _ = try controller.clear(request: req)
+        _ = try controller.clear(req)
     }
 }
